revert DRYing up validation: min length not required for slack integration
