+++ packages/back-end/src/models/EventWebhookModel.ts
@@ -2,10 +2,14 @@ import { randomUUID } from "crypto";
 import omit from "lodash/omit";
 import md5 from "md5";
 import mongoose from "mongoose";
-import { NotificationEventName } from "../events/base-types";
+import { z } from "zod";
+import {
+  NotificationEventName,
+  notificationEventNames,
+} from "../events/base-types";
 import { EventWebHookInterface } from "../../types/event-webhook";
 import { logger } from "../util/logger";
-import { validateNotificationEventNames } from "./validators/validateNotificationEventNames";
+import { errorStringFromZodResult } from "../util/validation";
 
 const eventWebHookSchema = new mongoose.Schema({
   id: {
@@ -37,7 +41,18 @@ const eventWebHookSchema = new mongoose.Schema({
     type: [String],
     required: true,
     validate: {
-      validator: validateNotificationEventNames,
+      validator(value: unknown) {
+        const zodSchema = z.array(z.enum(notificationEventNames)).min(1);
+
+        const result = zodSchema.safeParse(value);
+
+        if (!result.success) {
+          const errorString = errorStringFromZodResult(result);
+          logger.error(errorString, "Invalid Event name");
+        }
+
+        return result.success;
+      },
     },
   },
   url: {
